move profiler from common to server
profile is only used at the server side now
